+++ projects/angular-sdk/CHANGELOG.md
@@ -8,6 +8,7 @@
 ### Minor changes
 - Offer Banner Large:
     - Improved the display of the error message when there are no offers
+    - Transmission of the premium price when selecting an offer
 
 
 <a name="2.9.0"></a>